Migrate to server-side API for treasury data fetching
- Remove static export configuration from next.config.ts
- Create /api/treasury server-side route using viem and Etherscan API
- Update useTreasuryData hook to call API route instead of direct client calls
- Move ETHERSCAN_API_KEY to server-side environment variable
- Update .env.example with server-only API key configuration
- Clean up unused imports and dependencies
